virt/wsl: implement wsl-shell to nsenter into systemd ns
The user is now correctly dropped into the Skiff system when running "wsl.exe"
from the PowerShell command line.

+++ configs/virt/wsl/hooks/post.sh
@@ -33,10 +33,18 @@ for b in ${BUSYBOX_BINS[@]}; do
     ln -fs ./busybox ${PERSIST_DIR}/bin/${b}
 done
 
+# wsl-shell
+cp ./skiff-init/wsl-shell ${PERSIST_DIR}/bin/wsl-shell
+rm ${BOOT_DIR}/skiff-init/wsl-shell || true
+
 # create WSL configs
 touch ${PERSIST_DIR}/etc/fstab
 echo "root:x:0:root" > ${PERSIST_DIR}/etc/group
-echo "root:x:0:0:root:/root:/bin/sh" > ${PERSIST_DIR}/etc/passwd
+if [ -f ${PERSIST_DIR}/bin/wsl-shell ]; then
+    echo "root:x:0:0:root:/root:/bin/wsl-shell" > ${PERSIST_DIR}/etc/passwd
+else
+    echo "root:x:0:0:root:/root:/bin/sh" > ${PERSIST_DIR}/etc/passwd
+fi
 cp ${SKIFF_CURRENT_CONF_DIR}/resources/wsl.conf ${PERSIST_DIR}/etc/wsl.conf
 
 # WSL loads a .tar.gz
@@ -44,3 +52,5 @@ cd ${PERSIST_DIR}; find -print0 | LC_ALL=C sort -z | tar \
     --pax-option="exthdr.name=%d/PaxHeaders/%f,atime:=0,ctime:=0" \
     -cf ${SKIFF_IMAGE} --null --xattrs-include='*' \
     --no-recursion -T - --numeric-owner
+echo "Created $(basename ${SKIFF_IMAGE}), import it with:"
+echo "# wsl.exe --import SkiffOS C:\SkiffOS skiffos.tar.gz"
